fix: 추가문의 본문 즉시 업데이트 안되는 오류 해결
