--- src/hooks/followup/followupApi.ts
@@ -1,29 +1,0 @@
-import { useMutation } from "@tanstack/react-query";
-
-import {
-  PostFollowupsRequest,
-  PutFollowupsRequest,
-} from "@/types/followup/followupApi";
-
-import { postFollowups, putFollowups } from "@/apis/followup/followupApi";
-
-export const useFollowupApi = (inquiryId: number) => {
-  const postFollowupsMutation = useMutation({
-    mutationFn: (data: PostFollowupsRequest) => postFollowups(inquiryId, data),
-  });
-
-  const putFollowupsMutation = useMutation({
-    mutationFn: ({
-      followupId,
-      data,
-    }: {
-      followupId: number;
-      data: PutFollowupsRequest;
-    }) => putFollowups(followupId, data),
-  });
-
-  return {
-    postFollowupsMutation,
-    putFollowupsMutation,
-  };
-};
